requirements.txt
flask
spacy
firebase-admin
PyPDF2
python-docx
scikit-learn
